# Convert malloc to calloc in bigarray