--- a/Changes
+++ b/Changes
@@ -77,6 +77,9 @@ Working version
   areas in the page table, subsumed by the new code fragment management API
   (Xavier Leroy, review by Jacques-Henri Jourdan)
 
+- #9725: Zero-initialize bigarrays in caml_ba_alloc
+  (Stefan Muenzel, review by ???)
+
 ### Code generation and optimizations:
 
 - #9620: Limit the number of parameters for an uncurried or untupled