Currently, bigarrays are not zero-initialized, which sometimes leads to bigarrays that contain memory that was previously used by the program.
Since this is a source of bugs, we change the bigarray creation to use "calloc" instead, which zero-initializes memory. calloc also does the overflow-free multiplication check for us -- so we eliminate this as well.