@xavierleroy 2020-06-30
This changes the cost of creating a bigarray from constant time to O(n), n being possibly quite large because these are BIGarrays.  Do we really want that?