@smuenzel-js 2020-06-30
I think we expect the user not to allocate very many big bigarrays during a run of the program, because of the following:
1) Writing useful info to the bigarrays will be O(n) anyway (I'm ignoring using them in a sparse manner for this particular point)
2) It is already worst-case O(n), because when if there isn't already a piece of memory available, the OS needs to provision memory. Yes, It can do that lazily, but if each page is touched, it will definitely need to both add the entry to the page table, and initialize the page to zero.
3) Large allocations should come from the OS directly anyway, and calloc seems to know that these are already zero-initialized.

I tried to verify what I just said with some benchmarks in c code, just running a single allocation of size 64GB.
I think the amusing result is that it doesn't actually matter,
malloc takes around 27us, calloc is actually faster, and takes 5us.
If you touch all the memory (i touched 1 byte every 1KB), it takes much much longer, 30s for malloc, 30s for calloc.

Hardware was an E5-2667 (that's an 8-year-old cpu!) with 128GB memory.