@dbuenzli 2020-06-30
@smuenzel-js could you please update the docs [here](https://github.com/ocaml/ocaml/blob/33b9b16dd75ae8cd208a139121a98a0e8783c99d/stdlib/bigarray.mli#L294-L295) to reflect what the PR proposes.