@avsm 2020-06-30
The only way I know that `calloc` will be faster than `malloc+memset` is that the `calloc` will reuse the zero page from the kernel.  This only delays the cost of accessing the page until later, as it will trigger a page fault when you write something into it to demand-map a fresh page in to replace the zero page.  So I'm not sure the microbenchmark from C is much use without seeing the specifics of the benchmark, what system you ran it on, and so on.

MirageOS has plenty of cases where we allocate a 4K page to a bigarray and then only write to a small portion of it (e.g. for IO zerocopy to Xen).  But we already have a [custom allocator](https://github.com/mirage/io-page/blob/master/lib/io_page.mli) that uses posix_memalign to give us page aligned Bigarrays, so the behaviour proposed in this PR wouldn't adversely affect us.

I guess the decision on this PR hinges on whether the undefined contents of `malloc` are actively harmful to OCaml code in terms of introducing new bugs.  Since these areas of memory are for data only, never code, it comes down to whether applications are being developed where users _believe_ that Bigarray will be zero-filled (which may be true when the executable first starts, as it will be getting zero'ed pages), and then subsequently when the process starts recycling pages they discover that the pages are recycled.  What sorts of bugs have you seen from this in the wild, @smuenzel-js?

Note that the system allocator is responsible for making sure that recycled malloc/free pages do not contain sensitive data. Many libcs can be configured to zero out new allocated memory (see e.g. [jemalloc's `opt.zero`](https://www.freebsd.org/cgi/man.cgi?jemalloc(3)), and this policy will work across all C bindings, not just Bigarrays.  OpenBSD has [*freezero(3)*](https://man.openbsd.org/free.3) which explicitly discards pages with sensitive data that should never be recycled.