@xavierleroy 2020-06-30
> I tried to verify what I just said with some benchmarks in c code, just running a single allocation of size 64GB. I think the amusing result is that it doesn't actually matter, malloc takes around 27us, calloc is actually faster, and takes 5us.

That `calloc` doesn't need to zero memory just obtained from the kernel is an interesting idea.  Using the silly test below I can see this working under macOS, but not under Linux (Ubuntu 20.04), where `calloc` takes time very much linear in the size of the allocated data.  So I would not rely on the OS and C standard library "doing the right thing" by themselves.

```
#include <stdlib.h>

int main(int argc, char ** argv)
{
  size_t sz;
  char * p;

  sz = (size_t) atoi(argv[1]) * 1024 * 1024 * 1024;
  p = calloc(sz, 1);
  return 0;
}

/*
Ubuntu 20.04   8 Gb    16 Gb    24 Gb

Ryzen 3700X    1.5s    3.0s     4.5s
Xeon W-2145    2.7s    5.4s     8.0s
*/
```