@xavierleroy 2020-06-30
> It is already worst-case O(n), because when if there isn't already a piece of memory available, the OS needs to provision memory.

Using `malloc` instead of `calloc` in my silly test, execution time (Ubuntu 20.04)  is under 1 ms regardless of the size.