@smuenzel-js 2020-07-01
Before we get too sidetracked measuring performance, I think I want to repeat my claim that linear performance for caml_ba_alloc doesn't matter.
Any program that actually uses the memory allocated will pay a linear penalty anyway -- the main penalty being initializing and mapping pages.
So the only programs that will have issue with the performance need to meet all these conditions:
1) Use the memory allocated in a sparse manner
2) Don't care about the non-written-to parts of the sparse memory
3) Have a runtime that is short compared to the cost of initializing the memory.
Do we know of such a program? My understanding of the use-cases of sparse memory generally require the non-written-to memory to be initialized to zero (because otherwise, you'd use a different data structure -- the design of a program meeting 1,2 and 3 seems really weird).

Now, regarding the errors that we observed in the wild:
I know of two different instances of this, in both cases, the code assumed that the array would be initialized to zero. The first case created a packet that would be sent to a remote service, and only wrote to the fields that it actually cared about setting.
The second piece of code initialized a static table by performing calculations on the table itself, and it also assumed that it would be initialized to zero.
The real problem is that we wrote tests for both of these, and the tests pass! This is due to the testing environment being isolated from other code, so the probability of reusing memory was very low -- it was always fresh memory, initialized to zero.
We only noticed when random tests elsewhere in the codebase (that ran more, other code) started failing. The problems were pretty difficult to track down, since it's hard to trace the involvement of the bigarray in the complex pieces of code that are required to produce a test failure.
There is a realistic chance that tests would not catch this, and that we'd only notice when a long running program in production exhibits symptoms -- and again, it would be hard to track down.

I think users (including me!) are being misled by typical conventions of ocaml that require all variables to be initialized before being used (e.g. Array.create). That bigarray is different seems like a usability issue -- I think having to remember special cases that only apply to a small subset of code makes writing error-free code much more difficult.


See this for my code: https://gist.github.com/smuenzel-js/e06a5b00af99bc444f4ef2a5e3eae37c
I tested with glibc 2.17 on kernel 3.10.0, Intel(R) Xeon(R) CPU E5-2667 v2 @ 3.30GHz

Without touching memory:

```
size   malloc   calloc
 1GB  25.9us  3.5us
 2GB  26.1us  3.6us
 4GB  25.0us  3.6us
 8GB  47.8us  4.0us
16GB  35.2us  3.5us
32GB  24.8us  3.6us
64GB  26.8us  3.9us
96GB  31.8us  4.0us
```

With touching memory

```
size   malloc   calloc
 1GB  0.39s  0.40s
 2GB  0.83s  0.84s
 4GB  1.61s  1.60s
 8GB  3.13s  3.26s
16GB  6.27s  6.34s
32GB  12.5s  12.8s
64GB  26.0s  32.5s
96GB  41.4s  38.4s
```