@bluddy 2020-07-01
This seems like a good candidate for adding to the API. 

Would it make sense to have a `Bigarray.make` which uses `calloc`, a `Bigarray.make_unsafe` which doesn't initialize the value (i.e. `malloc`) for performance, and then to deprecate `Bigarray.create` (consistent with the other libraries)?